requirements.txt
Flask
gspread
oauth2client
tefas-crawler
pandas
python-dotenv
